% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_source_mean.R
\name{calc_source_mean}
\alias{calc_source_mean}
\title{FUNCTION_TITLE}
\usage{
calc_source_mean(id_col, corrected_data_col, data)
}
\arguments{
\item{id_col}{PARAM_DESCRIPTION}

\item{corrected_data_col}{PARAM_DESCRIPTION}

\item{data}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
