% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sum_v.R
\name{calc_sum_v}
\alias{calc_sum_v}
\title{FUNCTION_TITLE}
\usage{
calc_sum_v(df)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
   df <- DATASET
   df <- label_TS_Nitrogen(df)
   df <- calc_sum_v(df)
 }
}
}
\seealso{
\code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{geom_point}},\code{\link[ggplot2]{aes}},\code{\link[ggplot2]{ggtheme}}
}
