% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_TP.R
\name{calc_TP}
\alias{calc_TP}
\title{FUNCTION_TITLE}
\usage{
calc_TP(
  df,
  AA_src = "Phe",
  AA_trp = "Glu",
  Beta = 3.3,
  Beta_SD = 1.8,
  eq_TDF_n = 1,
  TDF1 = 7.6,
  TDF1_SD = 1.4,
  TDF2 = 5.5,
  TDF2_SD = 0
)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{AA_src}{PARAM_DESCRIPTION, Default: 'Phe'}

\item{AA_trp}{PARAM_DESCRIPTION, Default: 'Glu'}

\item{Beta}{PARAM_DESCRIPTION, Default: 3.3}

\item{Beta_SD}{PARAM_DESCRIPTION, Default: 1.8}

\item{eq_TDF_n}{PARAM_DESCRIPTION, Default: 1}

\item{TDF1}{PARAM_DESCRIPTION, Default: 7.6}

\item{TDF1_SD}{PARAM_DESCRIPTION, Default: 1.4}

\item{TDF2}{PARAM_DESCRIPTION, Default: 5.5}

\item{TDF2_SD}{PARAM_DESCRIPTION, Default: 0}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
