% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_TS_Nitrogen.R
\name{label_TS_Nitrogen}
\alias{label_TS_Nitrogen}
\title{label_TS_Nitrogen}
\usage{
label_TS_Nitrogen(df)
}
\arguments{
\item{df}{data frame formatted as "DATASET"}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
factors your samples as tropic and source and plots to check data
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 data("DATASET")
 label_TS_Nitrogen(df = DATASET)
 }
}
}
\seealso{
\code{\link[pacman]{p_load}}
}
