% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_AA.R
\name{pull_AA}
\alias{pull_AA}
\title{FUNCTION_TITLE}
\usage{
pull_AA(df, AA_nm)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{AA_nm}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
